requirements.txt
pystray==0.19.5
Pillow==12.0.0
